Move models out of helper for Sqlite
